+++ tests/router/path.test.tsx
@@ -3,7 +3,7 @@ import "../env.js";
 import { strictEqual } from "node:assert";
 import test from "node:test";
 
-import { join, normalize, trimBase } from "@mxjp/gluon/router";
+import { join, normalize, relative } from "@mxjp/gluon/router";
 
 await test("router/path", async ctx => {
 	await ctx.test("normalize", () => {
@@ -58,24 +58,27 @@ await test("router/path", async ctx => {
 		strictEqual(join("/foo/", "/bar/"), "/foo/bar/");
 	});
 
-	await ctx.test("trimBase", () => {
-		strictEqual(trimBase("", ""), "");
-		strictEqual(trimBase("", "/"), "");
-		strictEqual(trimBase("/", ""), "");
+	await ctx.test("relative", () => {
+		strictEqual(relative("", ""), "");
+		strictEqual(relative("", "/"), "");
+		strictEqual(relative("/", ""), "");
 
-		strictEqual(trimBase("/", "foo"), "/foo");
-		strictEqual(trimBase("", "foo/bar"), "/foo/bar");
+		strictEqual(relative("/", "foo"), "/foo");
+		strictEqual(relative("", "foo/bar"), "/foo/bar");
 
-		strictEqual(trimBase("foo", "foo/bar"), "/bar");
-		strictEqual(trimBase("foo", "/foo/bar"), "/bar");
-		strictEqual(trimBase("/foo", "foo/bar"), "/bar");
-		strictEqual(trimBase("foo", "foo/bar/"), "/bar/");
-		strictEqual(trimBase("foo/", "foo/bar"), "/bar");
-		strictEqual(trimBase("foo/bar", "foo/bar/baz/boo"), "/baz/boo");
+		strictEqual(relative("foo", "foo/bar"), "/bar");
+		strictEqual(relative("foo", "/foo/bar"), "/bar");
+		strictEqual(relative("/foo", "foo/bar"), "/bar");
+		strictEqual(relative("foo", "foo/bar/"), "/bar/");
+		strictEqual(relative("foo/", "foo/bar"), "/bar");
 
-		strictEqual(trimBase("foo", "bar"), undefined);
-		strictEqual(trimBase("foo/", "bar"), undefined);
-		strictEqual(trimBase("foo", "foobar"), undefined);
-		strictEqual(trimBase("foo/", "foobar"), undefined);
+		strictEqual(relative("foo/bar", "foo/baz/boo"), "/../baz/boo");
+		strictEqual(relative("foo/bar", "foo/bar/baz/boo"), "/baz/boo");
+
+		strictEqual(relative("foo", "bar"), "/../bar");
+		strictEqual(relative("foo/", "bar"), "/../bar");
+		strictEqual(relative("foo", "foobar"), "/../foobar");
+		strictEqual(relative("foo/", "foobar/"), "/../foobar/");
+		strictEqual(relative("foo/bar", "baz/boo"), "/../../baz/boo");
 	});
 });